# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rent Contract API",
        "description": "API documentation for Rent Contract Service",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
        "/api/v1/contracts": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Get all contracts",
                "operationId": "ed99861593a025fe4bb15a5215d1132d",
                "responses": {
                    "200": {
                        "description": "Contracts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessCollectionResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Create a new contract",
                "operationId": "4f4bae3a5a6c17f01a25a8a68f3b6236",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContractRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contract created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessSingleResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/contracts/{id}": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Get a specific contract",
                "operationId": "75713383345080823820012ecd10dc23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of the contract",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contract retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessSingleResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contract not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Update a contract",
                "operationId": "da99b166039d0cea7a0086d6cf61d27b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of the contract",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContractRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contract updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessSingleResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contract not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Delete a contract",
                "operationId": "4de90671104f83b726ea85ca5ce73b9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of the contract",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contract deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SuccessSingleResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contract not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "TenantObject": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "email": {
                        "type": "string",
                        "example": "budi@example.com"
                    }
                },
                "type": "object"
            },
            "ContractResource": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "tenant_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "listing_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440002"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "example": "DRAFT",
                        "enum": [
                            "DRAFT",
                            "ACTIVE",
                            "EXPIRED",
                            "TERMINATED"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "tenant": {
                        "$ref": "#/components/schemas/TenantObject"
                    }
                },
                "type": "object"
            },
            "ContractRequest": {
                "required": [
                    "tenant_id",
                    "listing_id",
                    "start_date",
                    "end_date"
                ],
                "properties": {
                    "tenant_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "listing_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440002"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "string",
                        "example": "DRAFT",
                        "enum": [
                            "DRAFT",
                            "ACTIVE",
                            "EXPIRED",
                            "TERMINATED"
                        ]
                    }
                },
                "type": "object"
            },
            "ApiMeta": {
                "properties": {
                    "service_name": {
                        "type": "string",
                        "example": "Rent-Contract-Service"
                    },
                    "api_version": {
                        "type": "string",
                        "example": "v1"
                    }
                },
                "type": "object"
            },
            "SuccessCollectionResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Data retrieved successfully"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ContractResource"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/ApiMeta"
                    }
                },
                "type": "object"
            },
            "SuccessSingleResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Data retrieved successfully"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ContractResource"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/ApiMeta"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unable to process request"
                    },
                    "data": {
                        "example": null,
                        "nullable": true
                    },
                    "meta": {
                        "$ref": "#/components/schemas/ApiMeta"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiKeyAuth": {
                "type": "apiKey",
                "description": "Input API Key here.",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
...
}